--- react-materials/scaffolds/ice-asset-management-admin/src/pages/CostManage/components/CostTrend/CostTrend.scss
@@ -1,23 +1,0 @@
-@charset "UTF-8";
-
-.cost-trend {
-  .next-tabs {
-    overflow: visible;
-  }
-  .next-tabs-bar {
-    position: absolute;
-    top: -48px;
-    right: 0;
-  }
-  .next-tabs-text > .next-tabs-bar .next-tabs-tab:after {
-    width: 0;
-  }
-}
-
-@media screen and (max-width: 720px) {
-  .cost-trend {
-    .next-tabs-small .next-tabs-tab-inner {
-      padding: 0 5px;
-    }
-  }
-}
