+++ react-materials/scaffolds/ice-asset-management-admin/src/pages/CostManage/components/CostTrend/index.scss
@@ -1,6 +1,6 @@
-// @charset "UTF-8";
+@charset "UTF-8";
 
-// .trend-analysis {
+// .cost-trend {
 //   .next-tabs {
 //     overflow: visible;
 //   }
@@ -15,7 +15,7 @@
 // }
 
 // @media screen and (max-width: 720px) {
-//   .trend-analysis {
+//   .cost-trend {
 //     .next-tabs-small .next-tabs-tab-inner {
 //       padding: 0 5px;
 //     }
