+++ react-materials/scaffolds/ice-asset-management-admin/src/pages/CostManage/components/DepartmentCost/index.scss
@@ -1,23 +1,23 @@
 @charset "UTF-8";
 
 .department-cost {
-  .next-tabs {
-    overflow: visible;
-  }
-  .next-tabs-bar {
-    position: absolute;
-    top: -48px;
-    right: 0;
-  }
-  .next-tabs-text > .next-tabs-bar .next-tabs-tab:after {
-    width: 0;
-  }
+  // .next-tabs {
+  //   overflow: visible;
+  // }
+  // .next-tabs-bar {
+  //   position: absolute;
+  //   top: -48px;
+  //   right: 0;
+  // }
+  // .next-tabs-text > .next-tabs-bar .next-tabs-tab:after {
+  //   width: 0;
+  // }
 }
 
-@media screen and (max-width: 720px) {
-  .department-cost {
-    .next-tabs-small .next-tabs-tab-inner {
-      padding: 0 5px;
-    }
-  }
-}
+// @media screen and (max-width: 720px) {
+//   .department-cost {
+//     .next-tabs-small .next-tabs-tab-inner {
+//       padding: 0 5px;
+//     }
+//   }
+// }
