+++ react-materials/scaffolds/ice-asset-management-admin/src/pages/CostManage/index.jsx
@@ -4,17 +4,6 @@ import CostTrend from './components/CostTrend';
 import DepartmentCost from './components/DepartmentCost';
 
 export default class CostManage extends Component {
-  static displayName = 'CostManage';
-
-  static propTypes = {};
-
-  static defaultProps = {};
-
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
   render() {
     return (
       <div>
