+++ react-materials/scaffolds/ice-asset-management-admin/src/pages/Home/components/BudgetManage/Calculate.jsx
@@ -122,7 +122,7 @@ export default class Calculate extends Component {
 
 const styles = {
   container: {
-    margin: '10px 0',
+    padding: '0 20px 20px',
   },
   link: {
     margin: '0 5px',
