+++ react-materials/scaffolds/ice-asset-management-admin/src/pages/Home/components/Overview/index.jsx
@@ -84,6 +84,7 @@ export default class Overview extends Component {
     );
   }
 }
+
 const styles = {
   alignRight: {
     textAlign: 'right',
