+++ react-materials/scaffolds/ice-asset-management-admin/src/pages/Home/components/YearsAnalysis/index.scss
@@ -1,23 +1,23 @@
-@charset "UTF-8";
+// @charset "UTF-8";
 
-.years-analysis {
-  .next-tabs {
-    overflow: visible;
-  }
-  .next-tabs-bar {
-    position: absolute;
-    top: -48px;
-    right: 0;
-  }
-  .next-tabs-text > .next-tabs-bar .next-tabs-tab:after {
-    width: 0;
-  }
-}
+// .years-analysis {
+//   .next-tabs {
+//     overflow: visible;
+//   }
+//   .next-tabs-bar {
+//     position: absolute;
+//     top: -48px;
+//     right: 0;
+//   }
+//   .next-tabs-text > .next-tabs-bar .next-tabs-tab:after {
+//     width: 0;
+//   }
+// }
 
-@media screen and (max-width: 720px) {
-  .years-analysis {
-    .next-tabs-small .next-tabs-tab-inner {
-      padding: 0 5px;
-    }
-  }
-}
+// @media screen and (max-width: 720px) {
+//   .years-analysis {
+//     .next-tabs-small .next-tabs-tab-inner {
+//       padding: 0 5px;
+//     }
+//   }
+// }
